requirements.txt
torch>=2.5.0
pillow>=11.0.0
setuptools>=75.0.0
urllib3>=2.2.3
wheel>=0.44.0
torchvision
flask
requests
